+++ src/widgets/Multiple.h
@@ -35,7 +35,7 @@ public:
 
     uint8_t inputSize()  const override { return 1; }
     uint8_t outputSize() const override { return 0; }
-    uint8_t serializeStrings(uint8_t* buf) const override;
+    uint16_t serializeStrings(uint8_t* buf) const override;
     void serializeInput(uint8_t* buf)          const override;
     void serializeOutput(uint8_t*)           const override {}
     void deserializeInput(const uint8_t* buf)      override;
