+++ src/widgets/Widget.cpp
@@ -92,14 +92,14 @@ void RadioKit_Widget::_registerSelf() {
     _pushDeferred(this);
 }
 
-uint8_t RadioKit_Widget::serializeStrings(uint8_t* buf) const {
+uint16_t RadioKit_Widget::serializeStrings(uint8_t* buf) const {
     uint8_t mask = 0;
     if (_label[0]   != '\0') mask |= RK_STR_LABEL;
     if (_icon[0]    != '\0') mask |= RK_STR_ICON;
     if (_onText[0]  != '\0') mask |= RK_STR_ONTEXT;
     if (_offText[0] != '\0') mask |= RK_STR_OFFTEXT;
 
-    uint8_t out = 0;
+    uint16_t out = 0;
     buf[out++] = mask;
 
     auto _writeStr = [&](const char* s) {
